requirements-dev.txt
-r requirements.txt
flake8
flake8-django
flake8-bandit
flake8-bugbear
pep8-naming
black
isort
pip-autoremove
pre-commit
pytest>=5.4.0
pytest-cov==2.11.1
pytest-django
Faker